move data to movie.json | more information about films

+++ homework/film_list/app/src/main/java/com/example/film_list/Movie.kt
@@ -1,5 +1,5 @@
 package com.example.film_list
 
-data class Movie (val name: String, val director: String, val year: Int, val rating: Int) {
+data class Movie (val name: String, val director: String, val year: Int, val rating: Float) {
 
 }
